+++ src/pushi/base/__init__.py
@@ -28,13 +28,13 @@ __license__ = "Apache License, Version 2.0"
 from . import apn
 from . import handler
 from . import mail
-from . import push_api
 from . import state
 from . import web
+from . import web_push
 
 from .apn import APNHandler
 from .handler import Handler
 from .mail import MailHandler
-from .push_api import PushAPIHandler
 from .state import AppState, State
 from .web import WebHandler
+from .web_push import WebPushHandler